Move test setup to separate file
